PROJETO ONE PIECE - with MongoDb-05

+++ src/controllers/paleta.controller.js
@@ -74,9 +74,14 @@ const updatePaletaController = async (req, res) => {
   res.send(updatedPaleta);
 };
 
-const deletePaletaController = (req, res) => {
+const deletePaletaController = async (req, res) => {
   const idParam = req.params.id;
-  paletasService.deletePaletaService(idParam);
+
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido.' });
+  }
+
+  await paletasService.deletePaletaService(idParam);
   res.send({ message: 'Paleta deletada com sucesso!' });
 };
